added unsigned number

+++ c++/include/CPU.hpp
@@ -20,7 +20,7 @@ private:
     int reg[REGMAX];
 
     int CU();
-    int ALU(int a, int b, int op, bool flag);
+    int8_t ALU(int8_t a, int8_t b, int8_t op, bool flag);
 
 public:
     CPU();
